requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-docx==1.1.0
openai==1.3.8
faiss-cpu==1.8.0
boto3==1.34.34
python-multipart==0.0.6
pydantic==2.5.2
tiktoken==0.5.2
numpy>=1.21.0
typing-extensions>=4.0.0
numpy<2.0.0,>=1.21.0 